Add link health check script (#205)
* Do not use whitespace char in folder name

It's causing problems for the upcoming health check script.

* Add link health check script

This script health checks all URLs in this repo. If it lives on a
separate branch Travis can be configured to run this script
periodically.

* Fix broken links

I found these broken links after running the health check script. I
found the replacement for all of them so none of them had to be removed.