Delete old docs structure 🫡
With the move to Docusaurus there's a new doc structure which we need to
use, unfortunately. A lot of git history will go away here which is
something that always saddens me. Many people has contributed and put
this information together over the years, for many many people to read
and help. I can't remember the exact number of people this website has
every day (was years since I could access the stats) but it was in the
thousands. Thanks everyone ❤️

The docs will now live in the (tada) docs/ dir which will be moved to
root.